requirements.txt
flask
flask-cors
gunicorn
firebase-admin
google-auth-oauthlib
google-api-python-client
opencv-python
numpy
face_recognition
python-dotenv